Work-in-progress commit: reorganizing
