+++ lib/darksky.rb
@@ -1,5 +1,4 @@
 require 'httparty'
-require 'slack-ruby-bot'
 
 class Darksky
   def self.darksky_info
@@ -30,14 +29,3 @@ class Darksky
     return info['daily']['data'].first['summary']
   end
 end
-
-class Slackbot < SlackRubyBot::Bot
-  command 'weather now' do |client, data, match|
-    client.say(text: Darksky.weather_now, channel: data.channel)
-  end
-
-  command 'weather tomorrow' do |client, data, match|
-    client.say(text: Darksky.weather_tomorrow, channel: data.channel)
-  end
-end
-Slackbot.run
